+++ src/max_div/_core/_cli/__init__.py
@@ -2,6 +2,7 @@ from ._cli import *  # noqa: F403
 from ._cmd_benchmark import *  # noqa: F403
 from ._cmd_benchmark_internal import *  # noqa: F403
 from ._cmd_benchmark_solver import *  # noqa: F403
+from ._cmd_benchmark_solver_feasibility import *  # noqa: F403
 from ._cmd_benchmark_solver_list_problems import *  # noqa: F403
 from ._cmd_benchmark_solver_presets import *  # noqa: F403
 from ._cmd_benchmark_solver_strategies import *  # noqa: F403